Large commit to resolve issues with the code generator.
How code is compiled:
  When a symbol is generated by the parser, it's size is set to either 4 or 8
(long/int) and its stack offset is set to 0.

  When the code generator is called. Every time an L-Value is encountered, it's
stack offset is set to it's size plus the previouly allocated L-Value's stack
offset. If there is no previously allocated L-Value, it's stack offset is
set to it's size.

  Since the stack offset variable is now handled by the code generator,
variables can now be defined in any order.

  Many more changes to the structure and documentation of the code generator
have occurred. Care was taken to ensure that it is as readable as possible.

  Have a look into 'test.c' to see all features the compiler supports in
action to base your own debugging off of.

Other fixes:
  instr.c:
      Returned instruction suffix now depends on the register size, not
    the variable size. This cleas up many cases where instead of having to
    increment the registerID by one, you can simply pass the registerID by
    itself.
      When passing a variable's size, decrement the size by 1 to ensure it
    follows the rules of the function.

  reg.c:
    Now contains debug lines with print when a register is allocated and
  deallocated.

  stack.c/stack.h:
    Dependincy to the code generator. The code generator no longer uses
this file and it as since been deleted.
  test.c:
    Updated to contain all features currently implemented in the
  compiler.
Large commit to resolve issues with the code generator.

How code is compiled:
  When a symbol is generated by the parser, it's size is set to either 4 or 8
(long/int) and its stack offset is set to 0.

  When the code generator is called. Every time an L-Value is encountered, it's
stack offset is set to it's size plus the previouly allocated L-Value's stack
offset. If there is no previously allocated L-Value, it's stack offset is
set to it's size.

  Since the stack offset variable is now handled by the code generator,
variables can now be defined in any order.

  Many more changes to the structure and documentation of the code generator
have occurred. Care was taken to ensure that it is as readable as possible.

  Have a look into 'test.c' to see all features the compiler supports in
action to base your own debugging off of.

Other fixes:
  instr.c:
      Returned instruction suffix now depends on the register size, not
    the variable size. This cleas up many cases where instead of having to
    increment the registerID by one, you can simply pass the registerID by
    itself.
      When passing a variable's size, decrement the size by 1 to ensure it
    follows the rules of the function.

  reg.c:
    Now contains debug lines with print when a register is allocated and
  deallocated.

  stack.c/stack.h:
    Dependincy to the code generator. The code generator no longer uses
this file and it as since been deleted.
  test.c:
    Updated to contain all features currently implemented in the
  compiler.